moved ConnectionModal to shared
